# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Shop API Document",
        "contact": {
            "email": "arshamdev2001@gmail.com"
        }
    },
    "host": "localhost:8000",
    "basePath": "/api/v1/",
    "paths": {
...
        "/token/refresh-token": {
            "post": {
                "description": "Create new access token from refresh token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tokens"
                ],
                "summary": "Get New AccessToken",
                "parameters": [
                    {
                        "description": "refresh token body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iarsham_shop-api_internal_dto.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/responses.RefreshTokenResponse"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InterServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/": {
            "get": {
                "description": "Retrieve user information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InterServerErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update user information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update User",
                "parameters": [
                    {
                        "description": "update user body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iarsham_shop-api_internal_dto.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InterServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/register-login": {
            "post": {
                "description": "Create user with firstname / lastname / phone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Register And Login By Phone",
                "parameters": [
                    {
                        "description": "register and login body",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_iarsham_shop-api_internal_dto.RegisterLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/responses.LoginOKResponse"
                        }
                    },
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/responses.RegisterOKResponse"
                        }
                    },
                    "500": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InterServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_iarsham_shop-api_internal_dto.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2OTY5NDE4MTcsInBob25lIjoiKzk4OTAyMTMxMjIyNCIsInVzZXJfaWQiOiI1In0.hzmZdfltaMDWaiTwO8IG1uPEyXOsu3JBs6giU2BDeMI"
                }
            }
        },
        "github_com_iarsham_shop-api_internal_dto.RegisterLoginRequest": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "maxLength": 75,
                    "minLength": 1,
                    "example": "James"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 75,
                    "minLength": 1,
                    "example": "Rodriguez"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 11,
                    "example": "+989021112299"
                }
            }
        },
...
        "github_com_iarsham_shop-api_internal_dto.UpdateUserRequest": {
            "type": "object",
            "required": [
                "first_name",
                "last_name"
            ],
            "properties": {
                "first_name": {
                    "type": "string",
                    "maxLength": 75,
                    "minLength": 1,
                    "example": "James"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 75,
                    "minLength": 1,
                    "example": "Rodriguez"
                }
            }
        },
...
        "github_com_iarsham_shop-api_internal_models.Comments": {
            "type": "object",
            "properties": {
                "UpdatedAt": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_iarsham_shop-api_internal_models.Likes"
                    }
                },
                "message": {
                    "type": "string"
                },
                "productsSlug": {
                    "type": "string"
                },
                "usersID": {
                    "type": "integer"
                }
            }
        },
        "github_com_iarsham_shop-api_internal_models.Likes": {
            "type": "object",
            "properties": {
                "commentsID": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "usersID": {
                    "type": "integer"
                }
            }
        },
...
        "responses.InterServerErrorResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "responses.LoginOKResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string",
                    "example": "Success, otp was sent"
                }
            }
        },
...
        "responses.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2OTY5NDE4MTcsInBob25lIjoiKzk4OTAyMTMxMjIyNCIsInVzZXJfaWQiOiI1In0.hzmZdfltaMDWaiTwO8IG1uPEyXOsu3JBs6giU2BDeMI"
                }
            }
        },
        "responses.RegisterOKResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string",
                    "example": "Success, otp was sent"
                }
            }
        },
...
        "responses.UserResponse": {
            "type": "object",
            "properties": {
                "UpdatedAt": {
                    "type": "string"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_iarsham_shop-api_internal_models.Comments"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "last_name": {
                    "type": "string"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_iarsham_shop-api_internal_models.Likes"
                    }
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}